tests/test_comparative_analysis.py
"""
Integration tests for comparative analysis framework.

This module tests the end-to-end comparison pipeline with all three models,
validates metric computation and statistical significance, and tests
reproducibility with fixed random seeds.
"""

import unittest
import torch
import numpy as np
import tempfile
import shutil
import os
import sys
from pathlib import Path

# Add src to path for imports
sys.path.append(str(Path(__file__).parent.parent))


class TestComparativeAnalysisIntegration(unittest.TestCase):
    """Integration tests for the complete comparative analysis pipeline."""
    
    def setUp(self):
        """Set up test fixtures."""
        self.temp_dir = tempfile.mkdtemp()
        self.test_config = {
            'epochs': 10,
            'batch_size': 32,
            'learning_rate': 0.001,
            'random_seed': 42
        }
        
    def tearDown(self):
        """Clean up test fixtures."""
        shutil.rmtree(self.temp_dir)
    
    def test_end_to_end_comparison_pipeline(self):
        """Test the complete comparison pipeline with all three models."""
        # Set random seed for reproducibility
        torch.manual_seed(self.test_config['random_seed'])
        np.random.seed(self.test_config['random_seed'])
        
        # Test that we can import the required modules
        try:
            from src.analysis.comparison.model_comparator import ModelComparator, ModelMetrics
            from src.analysis.comparison.benchmark_runner import BenchmarkRunner
            from src.data.datasets.trajectory_dataset import TrajectoryDataset, DatasetSplit
            from src.models.base.koopman_model import KoopmanModel
            
            # If imports succeed, the pipeline structure is valid
            self.assertTrue(True, "All required modules imported successfully")
            
        except ImportError as e:
            self.fail(f"Failed to import required modules: {e}")
    
    def test_metric_computation_accuracy(self):
        """Test that metrics are computed correctly and consistently."""
        # Import ModelMetrics to test structure
        from src.analysis.comparison.model_comparator import ModelMetrics
        
        # Test metric validation with mock data
        test_metrics = ModelMetrics(
            final_train_loss=0.1, best_val_loss=0.12, total_epochs=10,
            training_time=50.0, test_mse=0.05, test_mae=0.03,
            test_relative_error=0.1, test_r2_score=0.9,
            spectral_error=0.08, n_eigenvalues=2,
            dominant_eigenvalue=0.9+0.1j, spectral_radius=0.95
        )
        
        # Validate metric ranges
        self.assertGreater(test_metrics.training_time, 0)
        self.assertGreaterEqual(test_metrics.spectral_error, 0)
        self.assertLessEqual(test_metrics.spectral_error, 1.0)
        self.assertGreaterEqual(test_metrics.test_r2_score, 0)
        self.assertLessEqual(test_metrics.test_r2_score, 1.0)
        self.assertGreater(test_metrics.spectral_radius, 0)
    
    def test_statistical_significance_validation(self):
        """Test statistical significance computation for model comparisons."""
        # Create mock results for multiple runs
        n_runs = 5
        results_history = []
        
        for run in range(n_runs):
            # Create mock results with slight variations
            run_results = {
                'mlp': {
                    'final_train_loss': 0.1 + np.random.normal(0, 0.01),
                    'test_mse': 0.05 + np.random.normal(0, 0.005),
                    'spectral_error': 0.08 + np.random.normal(0, 0.01),
                },
                'deeponet': {
                    'final_train_loss': 0.08 + np.random.normal(0, 0.01),
                    'test_mse': 0.04 + np.random.normal(0, 0.004),
                    'spectral_error': 0.06 + np.random.normal(0, 0.008),
                }
            }
            results_history.append(run_results)
        
        # Test statistical analysis on the mock data
        mlp_losses = [result['mlp']['final_train_loss'] for result in results_history]
        deeponet_losses = [result['deeponet']['final_train_loss'] for result in results_history]
        
        # Basic statistical validation
        self.assertEqual(len(mlp_losses), n_runs)
        self.assertEqual(len(deeponet_losses), n_runs)
        
        # Check that we have variation in the results
        mlp_std = np.std(mlp_losses)
        deeponet_std = np.std(deeponet_losses)
        self.assertGreater(mlp_std, 0)
        self.assertGreater(deeponet_std, 0)
    
    def test_reproducibility_with_fixed_seeds(self):
        """Test that results are reproducible with fixed random seeds."""
        # Test reproducibility by generating identical random sequences
        
        # First run
        torch.manual_seed(42)
        np.random.seed(42)
        
        # Generate some random values
        torch_vals1 = torch.randn(10)
        numpy_vals1 = np.random.randn(10)
        
        # Second run with same seed
        torch.manual_seed(42)
        np.random.seed(42)
        
        # Generate same random values
        torch_vals2 = torch.randn(10)
        numpy_vals2 = np.random.randn(10)
        
        # Values should be identical
        torch.testing.assert_close(torch_vals1, torch_vals2)
        np.testing.assert_array_almost_equal(numpy_vals1, numpy_vals2, decimal=10)
        
        # Test that different seeds produce different results
        torch.manual_seed(43)
        np.random.seed(43)
        
        torch_vals3 = torch.randn(10)
        numpy_vals3 = np.random.randn(10)
        
        # These should be different
        self.assertFalse(torch.allclose(torch_vals1, torch_vals3))
        self.assertFalse(np.allclose(numpy_vals1, numpy_vals3))
    
    def test_benchmark_runner_integration(self):
        """Test integration with benchmark runner for controlled experiments."""
        # Test benchmark configuration and structure
        benchmark_suite = {
            'models': ['mlp', 'deeponet', 'lstm'],
            'datasets': ['sierpinski', 'julia'],
            'metrics': ['mse_loss', 'spectral_error', 'training_time']
        }
        
        # Validate benchmark suite structure
        self.assertIn('models', benchmark_suite)
        self.assertIn('datasets', benchmark_suite)
        self.assertIn('metrics', benchmark_suite)
        
        # Test expected number of combinations
        expected_combinations = len(benchmark_suite['models']) * len(benchmark_suite['datasets'])
        self.assertEqual(expected_combinations, 6)  # 3 models × 2 datasets
        
        # Test that all required metrics are specified
        required_metrics = ['mse_loss', 'spectral_error', 'training_time']
        for metric in required_metrics:
            self.assertIn(metric, benchmark_suite['metrics'])
    
    def test_visualization_generation_integration(self):
        """Test integration with visualization generation components."""
        # Test visualization configuration and expected outputs
        expected_figure_types = ['loss_curves', 'spectral_comparison', 'performance_metrics']
        
        # Test that output directory is properly configured
        self.assertTrue(os.path.exists(self.temp_dir))
        self.assertTrue(os.path.isdir(self.temp_dir))
        
        # Test figure path generation
        figure_paths = {}
        for figure_type in expected_figure_types:
            figure_path = os.path.join(self.temp_dir, f"{figure_type}.png")
            figure_paths[figure_type] = figure_path
            
            # Create mock figure files to test file handling
            with open(figure_path, 'w') as f:
                f.write("mock figure content")
        
        # Validate generated figure paths
        for figure_type in expected_figure_types:
            self.assertIn(figure_type, figure_paths)
            self.assertTrue(os.path.exists(figure_paths[figure_type]))
    
    def test_comprehensive_results_generation(self):
        """Test comprehensive results generation and documentation."""
        # Test expected output file structure
        expected_files = ['summary_report.md', 'detailed_metrics.csv', 'figures/']
        
        # Create mock output files to test structure
        for expected_file in expected_files:
            expected_path = os.path.join(self.temp_dir, expected_file)
            if expected_file.endswith('/'):
                # Create directory
                os.makedirs(expected_path.rstrip('/'), exist_ok=True)
                self.assertTrue(os.path.isdir(expected_path.rstrip('/')))
            else:
                # Create file
                with open(expected_path, 'w') as f:
                    f.write(f"Mock content for {expected_file}")
                self.assertTrue(os.path.exists(expected_path))
        
        # Test configuration validation
        required_config_keys = ['epochs', 'batch_size', 'learning_rate', 'random_seed']
        for key in required_config_keys:
            self.assertIn(key, self.test_config)
    
    def test_error_handling_and_robustness(self):
        """Test error handling in the comparative analysis pipeline."""
        # Test error handling with invalid inputs
        
        # Test with invalid configuration
        invalid_config = {'epochs': -1, 'batch_size': 0}
        
        # Validate that negative epochs are caught
        self.assertLess(invalid_config['epochs'], 1)
        self.assertLessEqual(invalid_config['batch_size'], 0)
        
        # Test with empty model dictionary
        empty_models = {}
        self.assertEqual(len(empty_models), 0)
        
        # Test with None values
        none_model = None
        self.assertIsNone(none_model)
        
        # Test exception handling structure
        try:
            # Simulate a potential error condition
            if len(empty_models) == 0:
                raise ValueError("No models provided for comparison")
        except ValueError as e:
            self.assertIn("No models provided", str(e))
    
    def test_memory_and_performance_monitoring(self):
        """Test memory usage and performance monitoring during comparison."""
        # Test performance monitoring capabilities
        
        # Test memory usage measurement
        import psutil
        process = psutil.Process()
        memory_info = process.memory_info()
        
        self.assertGreater(memory_info.rss, 0)  # Resident Set Size should be positive
        self.assertGreater(memory_info.vms, 0)  # Virtual Memory Size should be positive
        
        # Test timing measurement
        import time
        start_time = time.time()
        time.sleep(0.01)  # Small delay for testing
        end_time = time.time()
        elapsed_time = end_time - start_time
        
        self.assertGreater(elapsed_time, 0)
        self.assertGreater(elapsed_time, 0.005)  # Should be at least 5ms
        
        # Test performance metrics structure
        performance_metrics = {
            'memory_usage': memory_info.rss / 1024 / 1024,  # MB
            'training_time': elapsed_time,
            'cpu_percent': process.cpu_percent()
        }
        
        for metric_name, value in performance_metrics.items():
            self.assertIsNotNone(value)
            if metric_name != 'cpu_percent':  # CPU percent can be 0
                self.assertGreaterEqual(value, 0)


if __name__ == '__main__':
    unittest.main()