requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Deep learning framework
torch>=1.12.0
torchvision>=0.13.0

# Data manipulation and analysis
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration management
pyyaml>=6.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Progress bars and utilities
tqdm>=4.62.0

# Scientific plotting enhancements
scienceplots>=2.0.0

# Optional: GPU acceleration (uncomment if using CUDA)
# torch-audio>=0.12.0

# Development and testing (optional)
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Documentation (optional)
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0